[Place Page SEO] read text summaries from GCS directly (#3903)
This PR loads summary text on the place pages by making a read into the
GCS bucket in the place page endpoint, instead of loading them all to
the server app config on server load. This change means we won't have to
restart the prod server each time we make a change to the content of the
summaries.